code/requirements.txt
Django==4.2.5
mysqlclient==2.2.0
gunicorn==21.2.0
Pillow==9.5.0
django-extensions
django-crispy-forms
crispy-bootstrap4